chore: move permissions logic from settings_utils to permissions
